+++ error_messages.ts
@@ -1,0 +1,19 @@
+// Copyright 2018-2020 Gamebridge.ai authors. All rights reserved. MIT license.
+
+export const ERROR_MISSING_PROPERTIES_MAP =
+  "Unable to load serializer properties for the given context";
+
+export const ERROR_DUPLICATE_PROPERTY_KEY =
+  "This key name is already in use by another property, please use a different name";
+
+export const ERROR_DUPLICATE_SERIALIZE_KEY =
+  "This serialize key is already in use by another property, please use a different name";
+
+export const ERROR_SYMBOL_PROPERTY_NAME =
+  "The key name cannot be inferred from a symbol. A value for serializedName must be provided";
+
+export const ERROR_MISSING_STATIC_OR_VALUE_ON_POLYMORPHIC_SWITCH =
+  "Missing static property value or value in PolymorphicSwitch";
+
+export const ERROR_FAILED_TO_RESOLVE_POLYMORPHIC_CLASS =
+  "Failed to resolve polymorphic class";