+++ strategy/to_json/default_test.ts
@@ -1,0 +1,11 @@
+// Copyright 2018-2020 Gamebridge.ai authors. All rights reserved. MIT license.
+
+import { assertEquals, test } from "../../test_deps.ts";
+import { toJSONDefault } from "./default.ts";
+
+test({
+  name: "toJSONDefault passes the value unmodified",
+  fn() {
+    assertEquals(toJSONDefault("Hello world"), "Hello world");
+  },
+});